Split up the two HWIBridges properly
